requirements.txt
# PDF Processing
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.1

# Computer Vision & Layout Detection
opencv-python==4.8.1.78
layoutparser[ocr]==0.3.4
pytesseract==0.3.10

# SAM2 and Grounding-DINO
segment-anything-2==0.0.1
groundingdino-py==0.4.0
torch==2.1.0
torchvision==0.16.0

# Image Processing
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.66.1
requests==2.31.0

# Web Interface
Flask==2.3.3
Werkzeug==2.3.7
supervision>=0.16.0
inference-sdk
